+++ package.json
@@ -28,6 +28,7 @@
         "typescript": "^4.7.3"
     },
     "dependencies": {
+        "dotenv": "^16.0.1",
         "uuid": "^8.3.2"
     }
 }
